fix: resolve critical AML simulation issues in scheduling, transaction limits, and date handling
Re-enabled CASH-IN and CASH-OUT execution by applying configured normal/SAR intervals.
Fixed FanInTypology schedule handling to align with documented schedule_id semantics and safely handle empty originator sets.
Corrected transaction_limit accounting in TransactionRepository (removed invalid counter decrement).
Hardened AML typology edge cases (invalid ranges, empty arrays, and zero/negative-period scenarios).
Removed unsafe alertID narrowing cast in RandomTypology (long is now preserved).
Updated convert_logs.py so SAR event dates use the configured base_date instead of a fixed default.
Validation completed: Java tests passed (mvn test, 8/8) and Python tests passed (pytest, 20/20).

+++ src/main/java/amlsim/model/aml/RandomTypology.java
@@ -61,7 +61,7 @@ public class RandomTypology extends AMLTypology {
         Account bene = beneList.get(idx);
 
         TargetedTransactionAmount transactionAmount = new TargetedTransactionAmount(nextOrig.getBalance(), random);
-        makeTransaction(step, transactionAmount.doubleValue(), nextOrig, bene, isSAR, (int)alertID);  // Main account makes transactions to one of the neighbors
+        makeTransaction(step, transactionAmount.doubleValue(), nextOrig, bene, isSAR, alertID);  // Main account makes transactions to one of the neighbors
         nextOrig = bene;  // The next originator account is the previous beneficiary account
     }
 }
